requirements.txt
acme==0.31.0
aiohttp==3.6.2
asn1crypto==0.24.0
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
cached-property==1.5.1
certbot==0.31.0
certifi==2018.8.24
chardet==3.0.4
ConfigArgParse==0.13.0
configobj==5.0.6
cryptography==2.6.1
Deprecated==1.2.10
discord.py==1.4.0a2386+g2628949
docker==3.4.1
docker-compose==1.21.0
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
fail2ban==0.10.2
future==0.16.0
gitdb==4.0.5
github.py==0.5.0
github3.py==1.3.0
GitPython==3.1.7
httplib2==0.11.3
idna==2.6
josepy==1.1.0
jsonschema==2.6.0
jwcrypto==0.7
keyring==17.1.1
keyrings.alt==3.1.1
mock==2.0.0
multidict==4.7.6
numpy==1.19.1
panda==0.3.1
pandas==1.1.0
parsedatetime==2.4
pbr==4.2.0
psutil==5.7.0
pycrypto==2.6.1
pycurl==7.43.0.2
PyGithub==1.51
PyGObject==3.30.4
pyinotify==0.9.6
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.2.0
pyRFC3339==1.1
PySimpleSOAP==1.16.2
python-apt==1.8.4.1
python-dateutil==2.8.1
python-debian==0.1.35
python-debianbts==2.8.2
pytz==2019.1
pyxdg==0.25
PyYAML==3.13
reportbug===7.5.3-deb10u1
requests==2.21.0
requests-toolbelt==0.8.0
SecretStorage==2.3.1
six==1.12.0
smmap==3.0.4
soupsieve==2.0.1
systemd-python==234
texttable==1.6.0
timeago==1.0.13
ufw==0.36
uritemplate==3.0.1
urllib3==1.24.1
virtualenv==15.1.0
websocket-client==0.53.0
websockets==8.1
wrapt==1.12.1
yarl==1.1.1
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
